+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
 * Unicorn support (requires "preload_app true" in unicorn config file)
 * Fix for Thin detection - ensure it's actually running
+* Fixing name conflict btw Tracer#store and ActiveRecord::Store
 
 # 1.0.0